requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.15.2
matplotlib>=3.3
dj-database-url
